Created ts-definition for index-file.
