--- packages/bem-priv-class/index.js
@@ -1,2 +1,0 @@
-module.exports = require('./dist/index.js');
-module.exports.blockName = require('./dist/blockName.js');
