+++ packages/bem-priv-class/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "author": "notary <nkviz40k@gmail.com>",
   "description": "BEM-PRIV with ES6 classes and Typescript",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "build": "tsc",
     "build:test": "tsc -p test",
